+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to `laravel-mailcoach` will be documented in this file
 
 ## 4.11.0 - 2021-09-30
 
-- Use protected props for import subscribers action (#741)
-- Add an option to append tags when updating subscribers (#739)
+## What's fixed
 - Fix the ConditionAction breaking on different locales (#715)
 - Fix the subscribed landing view not being accessible without being logged in
+- The subject of transactional mail templates will now run replacers (#636)
+
+## What's new
+- Use protected props for import subscribers action (#741)
+- Add an option to append tags when updating subscribers (#739)
 
 ## 4.10.2 - 2021-09-24